requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
attrs==25.4.0
backoff==2.2.1
banks==2.2.0
bcrypt==5.0.0
beautifulsoup4==4.14.2
Brotli==1.1.0
build==1.3.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
chromadb==1.3.3
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.6.7
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
durationpy==0.10
executing==2.2.1
faiss-cpu==1.12.0
fastapi==0.121.0
ffmpy==0.6.4
filelock==3.20.0
filetype==1.2.0
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.10.0
google-auth==2.42.1
googleapis-common-protos==1.71.0
gradio==5.49.1
gradio_client==1.13.3
greenlet==3.2.4
griffe==1.14.0
groovy==0.1.2
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==1.0.1
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
jsonpickle==4.1.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kubernetes==34.1.0
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.14.7
llama-index-cli==0.5.3
llama-index-core==0.14.7
llama-index-embeddings-openai==0.5.1
llama-index-indices-managed-llama-cloud==0.9.4
llama-index-instrumentation==0.4.2
llama-index-llms-openai==0.6.7
llama-index-readers-file==0.5.4
llama-index-readers-llama-parse==0.5.1
llama-index-readers-whisper==0.2.1
llama-index-utils-workflow==0.5.0
llama-index-vector-stores-chroma==0.5.3
llama-index-vector-stores-faiss==0.5.1
llama-index-workflows==2.10.3
llama-parse==0.6.54
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib-inline==0.2.1
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.2
numpy==2.3.4
oauthlib==3.3.1
onnxruntime==1.23.2
openai==1.109.1
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
overrides==7.7.0
packaging==25.0
pandas==2.2.3
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.5.0
posthog==5.4.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pydantic==2.11.10
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pypdf==6.1.3
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
pyvis==0.3.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rpds-py==0.28.0
rsa==4.9.1
ruff==0.14.3
safehttpx==0.1.7
semantic-version==2.10.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
stack-data==0.6.3
starlette==0.49.3
striprtf==0.0.26
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.22.1
tomlkit==0.13.3
tqdm==4.67.1
traitlets==5.14.3
typer==0.20.0
typer-slim==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.2.14
websocket-client==1.9.0
websockets==15.0.1
wrapt==1.17.3
yarl==1.22.0
zipp==3.23.0
